ecommerce_project/requirements.txt
Django
graphene-django
uv
django-unfold